requirements.txt
flask==3.1.0
requests==2.32.3
beautifulsoup4==4.13.3
nltk==3.9.1
transformers==4.50.0
gtts==2.5.4
torch==2.6.0
gradio==5.22.0
keybert==0.9.0
scikit-learn==1.6.1
deep-translator==1.11.4
pydub==0.25.1
sentence-transformers==3.4.1
streamlit==1.43.2
python-dotenv==1.0.0
